requirements.txt
discord.py
wavelink
youtube-dl
PyNaCl
google-generativeai
urllib